+++ mycodo/inputs/atlas_rgb.py
@@ -189,6 +189,9 @@ class InputModule(AbstractInput):
             else:
                 self.atlas_sensor.query('O,PROX,0')
 
+            # Throw out first measurement of Atlas Scientific sensor, as it may be prone to error
+            self.get_measurement()
+
     def initialize_sensor(self):
         from mycodo.devices.atlas_scientific_ftdi import AtlasScientificFTDI
         from mycodo.devices.atlas_scientific_i2c import AtlasScientificI2C